NEW WAY TO ADD COMPONENTS TO /app
In the past, we had to run `npx ...`,
now we only need to run the server with 
`npm start`

start the webpack bundling to be in watch mode by
`npm run build`
and add new components into `src/components/` and use the in either the components or in the js files in `/src`. 

We should be good to go!